set cannonBubble's angle based on mouse's position
I have make sure the cannonbubble can be launched towards the position the mouse point at, but i still have tow issues, the first one is the program gets stuck for a while after every time i click on the screen and shoot a cannonbubble. I am not sure whether it is the code issue or its my computer issue. The second one is that the cannonbubble actually regard the mouse as an object too. So every time it hits the mouse, it will stop too.